old dangling changes
I should have committed these a long time ago.
Seems there is a lot of stuff here. I'm not sure if all of it
is actually functional, though, most seem to match what I have
already deployed.

Changes regarding Ruuvi-Collector are the bigger mystery.
I don't remember the reasoning for the branch change, as I use my
completely my own program, that seems to work fine.

Anyway. If something breaks, it's not a great loss. The
documentation is so bad, that it needs to be sorted out anyway.
I don't remember why I have this fork anyway...